+++ waves/wcore/urls.py
@@ -10,7 +10,7 @@ from waves.wcore.views.services import ServiceListView, ServiceDetailView
 urlpatterns = [
     url(r'^services/$', ServiceListView.as_view(), name='services_list'),
     url(r'^service/(?P<slug>[\w_-]+)/$', ServiceDetailView.as_view(), name='service_details'),
-    url(r'^service/(?P<slug>[\w_-]+)/create$', JobSubmissionView.as_view(), name='job_submission'),
+    url(r'^service/(?P<slug>[\w_-]+)/new', JobSubmissionView.as_view(), name='job_submission'),
     url(r'^jobs/(?P<slug>[\w-]+)/$', JobView.as_view(), name="job_details"),
     url(r'^jobs/inputs/(?P<slug>[\w-]+)/$', JobInputView.as_view(), name="job_input"),
     url(r'^jobs/outputs/(?P<slug>[\w-]+)/$', JobOutputView.as_view(), name="job_output"),